Rename : authentication 앱 이름 수정
- authentication 앱 이름을 authentications로 수정
