+++ mung_manager/authentications/apis/api_managers.py
@@ -2,7 +2,7 @@ from drf_spectacular.types import OpenApiTypes
 from drf_spectacular.utils import OpenApiResponse, extend_schema
 from rest_framework import status
 
-from mung_manager.authentication.apis.apis import KakaoLoginAPI
+from mung_manager.authentications.apis.apis import KakaoLoginAPI
 from mung_manager.common.base.api_managers import BaseAPIManager
 from mung_manager.schemas.errors.authentications import (
     ErrorAuthenticationUserInactiveSchema,
