+++ mung_manager/authentications/apis/apis.py
@@ -5,8 +5,8 @@ from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from mung_manager.authentication.containers import AuthenticationContainer
-from mung_manager.authentication.enums import UserProvider
+from mung_manager.authentications.containers import AuthenticationContainer
+from mung_manager.authentications.enums import UserProvider
 from mung_manager.common.base.serializers import BaseSerializer
 from mung_manager.errors.exceptions import AuthenticationFailedException
