+++ mung_manager/authentications/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class AuthenticationConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "mung_manager.authentication"
+    name = "mung_manager.authentications"
