+++ mung_manager/authentications/enums.py
