+++ mung_manager/authentications/selectors/abstracts.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import Optional
 
-from mung_manager.authentication.models import User
+from mung_manager.authentications.models import User
 from mung_manager.errors.exceptions import NotImplementedException
