+++ mung_manager/authentications/services/users.py
@@ -1,8 +1,8 @@
 from django.db import transaction
 
-from mung_manager.authentication.models import User
-from mung_manager.authentication.selectors.users import UserSelector
-from mung_manager.authentication.services.abstracts import AbstractUserService
+from mung_manager.authentications.models import User
+from mung_manager.authentications.selectors.users import UserSelector
+from mung_manager.authentications.services.abstracts import AbstractUserService
 from mung_manager.common.constants import SYSTEM_CODE
 from mung_manager.errors.exceptions import AuthenticationFailedException
